refactor[#405] : 임시저장 마킹 삭제 버튼 features/follow -> markings 변경
