+++ src/features/follow/api/index.ts
@@ -1,4 +1,3 @@
 export * from "./deleteFollowing";
 export * from "./postFollowing";
 export * from "./deleteFollower";
-export * from "./deleteTemporaryMarking";
