+++ src/features/follow/ui/index.ts
@@ -1,3 +1,2 @@
 export * from "./deleteFollowerButton";
 export * from "./followingToggle";
-export * from "./deleteTemporaryMarkingButton";
