+++ src/features/marking/ui/index.ts
@@ -4,3 +4,4 @@ export * from "./markingFilter";
 export * from "./markingLikeToggle";
 export * from "./markingBookmarkToggle";
 export * from "./tempMarkingFormModal";
+export * from "./deleteTemporaryMarkingButton";
