+++ src/mocks/handler.ts
@@ -6,8 +6,8 @@ import {
   LOGIN_END_POINT,
   SIGN_UP_END_POINT,
 } from "@/features/auth/constants";
-import { DeleteTemporaryMarkingRequest } from "@/features/follow/api";
 import { MY_MARKING_ENDPOINT } from "@/features/follow/constants";
+import { DeleteTemporaryMarkingRequest } from "@/features/marking/api";
 import { MARKING_END_POINT } from "@/features/marking/constants";
 import { PostChangeRegionRequest } from "@/features/setting/api";
 import type {
